Rename stopUpdatingStore to stopUpdateStore
