+++ lib/sync/stop-update-store.js
