+++ spec/api-spec.js
@@ -5,7 +5,7 @@ describe('API', () => {
     expect(typeof index.reducer).toEqual('function');
     expect(typeof index.updateStore).toEqual('function');
     expect(typeof index.updateStorageManager).toEqual('function');
-    expect(typeof index.stopUpdatingStore).toEqual('function');
+    expect(typeof index.stopUpdateStore).toEqual('function');
     expect(index.modelActions).toBeDefined();
     expect(index.collectionActions).toBeDefined();
   });
